code clean ups & validations added

+++ core-server/backend_server/backend_server/Handlers/V1/Reservations/Commands/Delete/Handler.cs
@@ -3,6 +3,7 @@ using backend_server.Models.DomainModels;
 using backend_server.Queries.Interfaces;
 using backend_server.Repositories.Interfaces;
 using MediatR;
+using static backend_server.Constants.ErrorConstant;
 
 namespace backend_server.Handlers.V1.Reservations.Commands.Delete;
 
@@ -19,8 +20,15 @@ public class Handler : IRequestHandler<Command, Response>
 
     public async Task<Response> Handle(Command command, CancellationToken cancellationToken)
     {
-        _ = await _reservationQuery.GetEntityById(command.Id)
+        var reservation = await _reservationQuery.GetEntityById(command.Id)
             ?? throw new NotFoundException(command.Id, nameof(Reservation));
+
+        TimeSpan difference = reservation.ReservationDate - DateTime.Now;
+
+        if (difference.TotalDays < 5)
+        {
+            throw new ValidationException(errorReason: ReservationError.InvalidTimePeriodReservationDeleteError);
+        }
 
         await _reservationRepository.Delete(command.Id);
